[Pipeline] - Goreleaser fueled CI releases (#49)
- Adds a pretty thorough goreleaser config h/t @edaniszewski
- Refactors out large chunks of the ci pipeline to be a little more
contained
- Leverages the future jenkins-agents containers for now to help in that
goal.

I haven't fully validated the tag build process, so i expect there may
be some hidden dragons in there. As well as the master build/push of
latest - but this should work as intended.

I remain on standby to support this PR and patch any bugs this may
introduce.